fix Issue #81:unix_socket use threadpool mode is not correct
when I test threadpool function,I found that the thread pool handles connection from the socket/admin port.
I think it is not correct,I think these two types of connections are administrator operations and should be distinguished from business connections,
otherwise the administrator may not be able to login when the business connection is abnormal (such as thread pool resource depletion,BUG, and other unknown problems).
In addition, administrators cannot create unrestricted connections(tdsql'agent will auto login tdsql by socket), so the extra_max_connections variable is added to control
how many connections an administrator can have at the same time, and the status counter for an administrator connection is added: Threads_extra_connected

briefly:

1.in thread pool mode,the unix socket/admin connection should not use threadpool

2.add variable: extra_max_connections to limit the socket connection/admin connection number.

3.add status: Threads_extra_connected to stat current's admin/socket connction

4.modify/add some mtr code

+++ sql/dd/impl/cache/shared_multi_map.h
@@ -265,7 +265,7 @@ class Shared_multi_map : public Multi_map_base<T> {
 
   bool pool_capacity_exceeded() const {
     mysql_mutex_assert_owner(&m_lock);
-    return m_element_pool.size() > max_connections;
+    return m_element_pool.size() > (size_t)max_connections;
   }
 
   /**
